refactor: move greetings code to separate package
